DARWIN: Yabai from Homebrew for a moment
I upgraded to Monterey, and didn't realize that Yabai didn't support it
yet.  The upstream author has made some progress towards supporting
Monterey, but it's not officially released.  Unfortunately, even if it
were officially released, there's more work to do to get it building
with Nix, because it requires XCode 13.1, which isn't in Nixpkgs yet.

So for now, I'm just installing this pre-release version with Homebrew
until we can build it with Nix.